🏗️ refactor ingredient feature

+++ src/ingredients/MW.SAXSAY.Ingredients.Domain/ValueObjects/IngredientId.cs
@@ -1,3 +1,3 @@
 namespace MW.SAXSAY.Ingredients.Domain.ValueObjects;
 
-public record IngredientId(int Value);
+public record IngredientId(long Value);
